chang color theme, make menu better, and buttons nicer.
Decided to change the theme of the website, looks better, Instead of a photo and random quote there, i watched some information on design and answered the main focus of why someone would be on my personal website, and that's to see my projects or to know more about me. So i answered that then immediately made a call to action along with previews. Have not added the links just yet, but soon.